refactor(gateway): split proxy, alerts, metering, and helpers into separate files
